+++ example/test/test_injector.dart
@@ -1,7 +1,7 @@
 import 'package:get_it/get_it.dart';
 import 'package:injectable/injectable.dart';
 
-import 'test_injector.iconfig.dart';
+import 'test_injector.config.dart';
 
 final getIt = GetIt.instance;