+++ injectable/lib/src/injectable_annotations.dart
@@ -1,114 +1,139 @@
+/// Marks a top-level function as an initializer function
+/// for configuring Get_it
 class InjectableInit {
-  // only files exist in provided directories will be processed
+  /// Only files exist in provided directories will be processed
   final List<String> generateForDir;
 
+  /// default constructor
   const InjectableInit({this.generateForDir = const ['lib', 'bin']})
-      : assert(generateForDir != null);
+      : assert(
+          generateForDir != null,
+        );
 }
 
-const injectableInit = const InjectableInit();
+/// const instance of [InjectableInit]
+/// with default arguments
+const injectableInit = InjectableInit();
 
+/// Marks a class as an injectable
+/// dependency and generates
 class Injectable {
-  // The type to bind your implementation to,
-  // typically an abstract class which is implemented by the
-  // annotated class.
+  /// The type to bind your implementation to,
+  /// typically, an abstract class which is implemented by the
+  /// annotated class.
   final Type as;
 
-  // an alternative way to pass env keys instead
-  // of annotating the element with @Environment
-  final String env;
+  /// an alternative way to pass env keys instead
+  /// of annotating the element with @Environment
+  final List<String> env;
 
+  /// default constructor
   const Injectable({this.as, this.env});
 }
 
-const injectable = const Injectable();
+/// const instance of [Injectable]
+/// with default arguments
+const injectable = Injectable();
 
 /// Classes annotated with @Singleton
-/// will generate registerSingleton func
+/// will generate registerSingleton function
 class Singleton extends Injectable {
+  /// passed to singlesReady property
+  /// in registerSingleton function
   final bool signalsReady;
+
+  /// passed to dependsOn property
+  /// in registerSingleton function
   final List<Type> dependsOn;
 
+  /// default constructor
   const Singleton({
     this.signalsReady,
     this.dependsOn,
     Type as,
-    String env,
+    List<String> env,
   }) : super(as: as, env: env);
 }
 
-const singleton = const Singleton();
+/// const instance of [Singleton]
+/// with default arguments
+const singleton = Singleton();
 
 /// Classes annotated with @LazySingleton
 /// will generate registerLazySingleton func
 class LazySingleton extends Injectable {
+  /// default constructor
   const LazySingleton({
     Type as,
-    String env,
+    List<String> env,
   }) : super(as: as, env: env);
 }
 
-const lazySingleton = const LazySingleton();
+/// const instance of [LazySingleton]
+/// with default arguments
+const lazySingleton = LazySingleton();
 
-// Used to annotate a constructor dependency
-// that's registered with an instance names;
+/// Used to register a dependency under a name
+/// instead of type also used to annotated
+/// named injected dependencies in constructors
 class Named {
-  // the name in which an instance is registered
+  /// The name in which an instance is registered
   final String name;
 
+  /// default constructor
   const Named(this.name) : type = null;
 
-  // instead of providing a literal name
-  // you can pass a type and it's name will be extracted
-  // in during generation
+  /// instead of providing a literal name
+  /// you can pass a type, and its name will be extracted
+  /// in during generation
   final Type type;
 
+  /// A named constrictor to extract name for type
   const Named.from(this.type) : name = null;
 }
 
-const named = const Named("");
+/// const instance of [Named]
+/// with default arguments
+const named = Named("");
 
-// used to annotate dependencies which are
-// registered under different environment
+/// Used to annotate dependencies which are
+/// registered under certain environments
 class Environment {
+  /// name of the environment
   final String name;
 
+  /// default constructor
   const Environment(this.name);
 
+  /// preset of common env name 'dev'
   static const dev = 'dev';
+
+  /// preset of common env name 'prod'
   static const prod = 'prod';
+
+  /// preset of common env name 'test'
   static const test = 'test';
 }
 
-const dev = const Environment(Environment.dev);
-const prod = const Environment(Environment.prod);
-const test = const Environment(Environment.test);
+/// preset instance of common env name
+const dev = Environment(Environment.dev);
 
-@Deprecated("Use @Injectable(as:...) or it's subs instead")
-// @Injectable(as: Type)
-// @Singleton(as: Type)
-// @LazySingleton(as: Type)
-class RegisterAs {
-  // The type to bind your implementation to,
-  // typically an abstract class which is implemented by the
-  // annotated class.
-  final Type abstractType;
+/// preset instance of common env name
+const prod = Environment(Environment.prod);
 
-  // an alternative way to pass env keys instead
-  // of annotating the element with @Environment
-  final String env;
+/// preset instance of common env name
+const test = Environment(Environment.test);
 
-  const RegisterAs(this.abstractType, {this.env});
-}
-
-// Marks a factory, a named constructor or a static create
-// function as an injectable constructor
-// if not added the default constructor will be used.
+/// Marks a factory, a named constructor or a static create
+/// function as an injectable constructor
+/// if not added the default constructor will be used.
 class FactoryMethod {
   const FactoryMethod._();
 }
 
-const factoryMethod = const FactoryMethod._();
+/// const instance of [FactoryMethod]
+/// with default values
+const factoryMethod = FactoryMethod._();
 
 /// Marks a constructor param as
 /// factoryParam so it can be passed
@@ -117,25 +142,28 @@ class FactoryParam {
   const FactoryParam._();
 }
 
-const factoryParam = const FactoryParam._();
+/// const instance of [FactoryParam]
+/// with default arguments
+const factoryParam = FactoryParam._();
 
-// marks a class as a register module where all
-// property accessors rerun types are considered factories
-// unless annotated with @singleton/lazySingleton.
-class RegisterModule {
-  const RegisterModule._();
+/// marks a class as a register module where all
+/// property accessors rerun types are considered factories
+/// unless annotated with @singleton/lazySingleton.
+class Module {
+  const Module._();
 }
 
-@Deprecated('Use module instead')
-const registerModule = const RegisterModule._();
+/// const instance of [Module]
+/// with default arguments
+const module = Module._();
 
-const module = const RegisterModule._();
-
-/// Futures annotated with [preResolv]
+/// Futures annotated with [preResolve]
 /// will be pre-awaited before they're
 /// registered inside of GetIt
 class PreResolve {
   const PreResolve._();
 }
 
-const preResolve = const PreResolve._();
+/// const instance of [PreResolve]
+/// with default arguments
+const preResolve = PreResolve._();