+++ injectable_generator/lib/dependency_resolver.dart
@@ -1,7 +1,7 @@
 import 'package:analyzer/dart/element/element.dart';
 import 'package:analyzer/dart/element/type.dart';
-import 'package:build/build.dart';
 import 'package:injectable/injectable.dart';
+import 'package:injectable_generator/import_resolver.dart';
 import 'package:injectable_generator/utils.dart';
 import 'package:source_gen/source_gen.dart';
 
@@ -13,7 +13,6 @@ const TypeChecker singletonChecker = TypeChecker.fromRuntime(Singleton);
 const TypeChecker injectableChecker = TypeChecker.fromRuntime(Injectable);
 
 const TypeChecker envChecker = TypeChecker.fromRuntime(Environment);
-const TypeChecker registerAsChecker = TypeChecker.fromRuntime(RegisterAs);
 const TypeChecker preResolveChecker = TypeChecker.fromRuntime(PreResolve);
 const TypeChecker factoryParamChecker = TypeChecker.fromRuntime(FactoryParam);
 const TypeChecker constructorChecker = TypeChecker.fromRuntime(FactoryMethod);
@@ -21,17 +20,14 @@ const TypeChecker constructorChecker = TypeChecker.fromRuntime(FactoryMethod);
 class DependencyResolver {
   Element _annotatedElement;
   final _dep = DependencyConfig();
-  final Resolver _resolver;
+  final ImportResolver _importResolver;
   final Map<String, String> _typeArgsMap = {};
 
-  DependencyResolver(this._resolver);
+  DependencyResolver(this._importResolver);
 
   Future<DependencyConfig> resolve(Element element) {
     _annotatedElement = element;
-    final import = getImport(element);
-    if (import != null) {
-      _dep.imports.add(import);
-    }
+    _dep.imports.add(_importResolver.resolve(element));
     return _resolveActualType(_annotatedElement);
   }
 
@@ -45,8 +41,7 @@ class DependencyResolver {
     if (returnType is ParameterizedType) {
       ClassElement element = returnType.element;
       for (int i = 0; i < element.typeParameters.length; i++) {
-        _typeArgsMap[element.typeParameters[i].name] =
-            returnType.typeArguments[i].getDisplayString();
+        _typeArgsMap[element.typeParameters[i].name] = returnType.typeArguments[i].getDisplayString();
       }
     }
 
@@ -56,11 +51,11 @@ class DependencyResolver {
       element: returnType.element,
     );
 
-    await _checkForParameterizedTypes(returnType);
+    _dep.imports.addAll(_importResolver.resolveAll(returnType));
 
     _dep.moduleName = moduleClazz.name;
     _dep.initializerName = executableElement.name;
-    _dep.imports.add(getImport(moduleClazz));
+    _dep.imports.add(_importResolver.resolve(moduleClazz));
 
     ExecutableElement executableModuleMember;
     if (executableElement is MethodElement) {
@@ -71,7 +66,7 @@ class DependencyResolver {
       } else {
         throwIf(
           executableElement.parameters.isNotEmpty,
-          'Abstract methods can not have injectable or factory paramters',
+          'Abstract methods can not have injectable or factory parameters',
           element: executableElement,
         );
       }
@@ -91,15 +86,15 @@ class DependencyResolver {
         clazz = returnType.element;
       }
     }
-    await _resolveAndAddImport(clazz);
+    _dep.imports.add(_importResolver.resolve(clazz));
     return _resolveActualType(clazz, typeName, executableModuleMember);
   }
 
   Future<DependencyConfig> _resolveActualType(
     ClassElement clazz, [
-    String typeName,
-    ExecutableElement executbaleModuleMemeber,
-  ]) async {
+        String typeName,
+        ExecutableElement excModuleMember,
+      ]) async {
     _dep.type = typeName ?? clazz.name;
     _dep.typeImpl = typeName ?? clazz.name;
 
@@ -118,25 +113,23 @@ class DependencyResolver {
         _dep.injectableType = InjectableType.lazySingleton;
       } else if (injectable.instanceOf(TypeChecker.fromRuntime(Singleton))) {
         _dep.injectableType = InjectableType.singleton;
-        _dep.signalsReady = injectable.peek('signalsReady')?.boolValue;
+        _dep.signalsReady = injectable
+            .peek('signalsReady')
+            ?.boolValue;
         _dep.dependsOn = injectable
             .peek('dependsOn')
             ?.listValue
             ?.map<String>((v) => v.toTypeValue().getDisplayString())
             ?.toList();
       }
-      abstractType = injectable.peek('as')?.typeValue;
-      inlineEnv = injectable.peek('env')?.stringValue;
-    }
-
-    if (abstractType == null) {
-      final registerAsAnnotation =
-          registerAsChecker.firstAnnotationOf(_annotatedElement);
-      if (registerAsAnnotation != null) {
-        final registerAsReader = ConstantReader(registerAsAnnotation);
-        abstractType = registerAsReader.peek('abstractType').typeValue;
-        inlineEnv = registerAsReader.peek('env')?.stringValue;
-      }
+      abstractType = injectable
+          .peek('as')
+          ?.typeValue;
+      inlineEnv = injectable
+          .peek('env')
+          ?.listValue
+          ?.map((e) => e.toStringValue())
+          ?.toList();
     }
 
     if (abstractType != null) {
@@ -153,15 +146,14 @@ class DependencyResolver {
       _dep.type = abstractSubtype.getDisplayString();
 
       _dep.typeImpl = clazz.name;
-      await _resolveAndAddImport(abstractSubtype.element);
-      await _checkForParameterizedTypes(abstractSubtype);
+      _dep.imports.addAll(_importResolver.resolveAll(abstractSubtype));
     }
 
-    _dep.environment = inlineEnv ??
+    _dep.environments = inlineEnv ??
         envChecker
-            .firstAnnotationOfExact(_annotatedElement, throwOnUnresolved: false)
-            ?.getField('name')
-            ?.toStringValue();
+            .annotationsOf(_annotatedElement)
+            ?.map((e) => e.getField('name')?.toStringValue())
+            ?.toList();
 
     _dep.preResolve = preResolveChecker.hasAnnotationOfExact(_annotatedElement);
 
@@ -177,18 +169,18 @@ class DependencyResolver {
       }
     }
 
-    ExecutableElement excutableInitilizer;
+    ExecutableElement executableInitilizer;
 
-    if (executbaleModuleMemeber != null) {
-      excutableInitilizer = executbaleModuleMemeber;
+    if (excModuleMember != null) {
+      executableInitilizer = excModuleMember;
     } else if (!_dep.isFromModule || _dep.isAbstract) {
       final possibleFactories = <ExecutableElement>[
         ...clazz.methods.where((m) => m.isStatic),
         ...clazz.constructors
       ];
 
-      excutableInitilizer = possibleFactories.firstWhere(
-          (m) => constructorChecker.hasAnnotationOfExact(m), orElse: () {
+      executableInitilizer = possibleFactories.firstWhere(
+              (m) => constructorChecker.hasAnnotationOfExact(m), orElse: () {
         throwIf(
           clazz.isAbstract,
           '''[${clazz.name}] is abstract and can not be registered directly!
@@ -197,20 +189,20 @@ class DependencyResolver {
         );
         return clazz.unnamedConstructor;
       });
-      _dep.isAsync = excutableInitilizer.returnType.isDartAsyncFuture;
+      _dep.isAsync = executableInitilizer.returnType.isDartAsyncFuture;
     }
 
-    if (excutableInitilizer != null) {
-      _dep.constructorName = excutableInitilizer.name;
-      for (ParameterElement param in excutableInitilizer.parameters) {
+    if (executableInitilizer != null) {
+      _dep.constructorName = executableInitilizer.name;
+      for (ParameterElement param in executableInitilizer.parameters) {
         final namedAnnotation = namedChecker.firstAnnotationOf(param);
         final instanceName = namedAnnotation
-                ?.getField('type')
-                ?.toTypeValue()
-                ?.getDisplayString() ??
+            ?.getField('type')
+            ?.toTypeValue()
+            ?.getDisplayString() ??
             namedAnnotation?.getField('name')?.toStringValue();
-        await _resolveAndAddImport(param.type.element);
-        await _checkForParameterizedTypes(param.type);
+
+        _dep.imports.addAll(_importResolver.resolveAll(param.type));
 
         var typeName = param.type.getDisplayString();
         if (param.type is TypeParameterType) {
@@ -241,7 +233,7 @@ class DependencyResolver {
 
       throwIf(
         _dep.isAbstract && factoryParamsCount != 0,
-        'Module dependecies with factory params must have custom initilaizers',
+        'Module dependencies with factory params must have custom initializers',
         element: clazz,
       );
 
@@ -260,41 +252,5 @@ class DependencyResolver {
     }
 
     return _dep;
-  }
-
-  Future<String> _resolveLibImport(Element element) async {
-    if (element == null ||
-        element.source == null ||
-        isCoreDartType(element.source)) {
-      return null;
-    }
-    //if element from a system library but not from dart:core
-    if (element.source.isInSystemLibrary) {
-      return getImport(element);
-    }
-    final assetId = await _resolver.assetIdForElement(element);
-
-    final toBeImported =
-        await _resolver.findLibraryByName(assetId.package) ?? element;
-    return getImport(toBeImported);
-  }
-
-  Future<void> _checkForParameterizedTypes(DartType paramType) async {
-    if (paramType is ParameterizedType) {
-      for (DartType type in paramType.typeArguments) {
-        await _checkForParameterizedTypes(type);
-        if (type.element.source != null) {
-          await _resolveAndAddImport(type.element);
-        }
-      }
-      ;
-    }
-  }
-
-  Future<void> _resolveAndAddImport(Element element) async {
-    final import = await _resolveLibImport(element);
-    if (import != null) {
-      _dep.imports.add(import);
-    }
   }
 }