+++ injectable_generator/lib/generator/factory_param_generator.dart
@@ -3,14 +3,12 @@ import 'package:injectable_generator/generator/register_func_generator.dart';
 
 class FactoryParamGenerator extends RegisterFuncGenerator {
   @override
-  String generate(DependencyConfig dep) {
+  String generate(DependencyConfig dep, {String prefix = '', String suffix = ''}) {
     final initializer = generateInitializer(dep);
 
     var asyncStr = dep.isAsync && !dep.preResolve ? 'Async' : '';
 
-    var typeArgs = dep.dependencies
-        .where((d) => d.isFactoryParam)
-        .fold<Map>(<String, String>{}, (all, b) => all..[b.paramName] = b.type);
+    var typeArgs = dep.dependencies.where((d) => d.isFactoryParam).fold<Map>(<String, String>{}, (all, b) => all..[b.paramName] = b.type);
     if (typeArgs.length < 2) {
       typeArgs['_'] = 'dynamic';
     }
@@ -18,10 +16,9 @@ class FactoryParamGenerator extends RegisterFuncGenerator {
     final argsDeclaration = '<${dep.type},${typeArgs.values.join(',')}>';
     final methodParams = typeArgs.keys.join(', ');
 
-    write(
-        "g.registerFactoryParam$asyncStr$argsDeclaration(($methodParams)=> $initializer");
+    write("${prefix}factoryParam$asyncStr$argsDeclaration(($methodParams)=> $initializer");
 
-    closeRegisterFunc(dep);
+    closeRegisterFunc(dep, suffix);
     return buffer.toString();
   }
 }